moved unit tests and updated changelog.md
